Move the property editor to ui.widgets
