+++ activity_browser/layouts/tabs/activity.py
@@ -1,5 +1,4 @@
 # -*- coding: utf-8 -*-
-from typing import Optional
 from peewee import DoesNotExist
 from PySide2 import QtCore, QtWidgets
 from PySide2.QtCore import Slot
@@ -7,8 +6,7 @@ from PySide2.QtCore import Slot
 from activity_browser import ab_settings, project_settings, signals
 from activity_browser.bwutils import commontasks as bc
 from activity_browser.mod import bw2data as bd
-from activity_browser.ui.tables.models.properties import PropertyModel
-from activity_browser.ui.tables.properties import PropertyTable
+from activity_browser.ui.widgets.property_editor import PropertyEditor
 
 from ...ui.icons import qicons
 from ...ui.style import style_activity_tab
@@ -371,80 +369,7 @@ class ActivityTab(QtWidgets.QWidget):
         ab_settings.write_settings()
 
     def open_properties(self):
-        """Opens the property editor for the current activity"""
-        editor = PropertyEditor(self.activity.get("properties"), self.read_only, self)
-        # Do not save the changes if the user pressed cancel
-        if editor.exec_() == editor.Accepted:
-            old_properties = self.activity.get("properties")
-            # Get the values modified by the user
-            updated_properties = editor.properties()
-            # Nothing to do
-            if not old_properties and not updated_properties:
-                return
-            # The user has deleted all properties
-            elif old_properties and not updated_properties:
-                del self.activity["properties"]
-            # There were no properties and the user created some
-            elif not old_properties and updated_properties:
-                self.activity["properties"] = updated_properties
-            else:
-                # Properties changed, merge the values, to avoid reordering
-                # the unmodified properties
-                for property in list(old_properties.keys()):
-                    if not property in updated_properties:
-                        del old_properties[property]
-                old_properties |= updated_properties
+        """Opens the property editor for the current """
+        # Do not save the changes if nothing changed
+        if PropertyEditor.edit_properties(self.activity, self.read_only, self):
             self.activity.save()
-
-
-
-class PropertyEditor(QtWidgets.QDialog):
-    """Property editor dialog"""
-
-    def __init__(self, properties: Optional[dict[str, float]], read_only: bool, parent: Optional[QtWidgets.QWidget] = None):
-        super().__init__(parent)
-        self.setWindowTitle("Property Editor")
-        self._data_model = PropertyModel(read_only)
-        if not read_only:
-            self._data_model.dataChanged.connect(self._handle_data_changed)
-        self._editor_table = PropertyTable(self._data_model)
-        self._editor_table.populate(properties)
-        self._save_button = QtWidgets.QPushButton("Save changes")
-        self._save_button.setEnabled(False)
-        self._save_button.clicked.connect(self.accept)
-        cancel_button = QtWidgets.QPushButton("Close" if read_only else "Cancel")
-        cancel_button.clicked.connect(self.reject)
-
-        # Prevent hitting enter in the table from closing the dialog
-        self._save_button.setAutoDefault(False)
-        cancel_button.setAutoDefault(False)
-        self._message_label = QtWidgets.QLabel()
-        if read_only:
-            self._message_label.setText("Read only")
-        else:
-            self._message_label.setText("No changes yet")
-    
-        layout = QtWidgets.QVBoxLayout()
-        layout.addWidget(self._editor_table)
-        button_layout = QtWidgets.QHBoxLayout()
-        button_layout.addWidget(self._save_button)
-        button_layout.addWidget(cancel_button)
-        layout.addLayout(button_layout)
-        layout.addWidget(self._message_label)
-        self.setLayout(layout)
-
-    def properties(self) -> dict[str, float]:
-        """Access method to get the result of the editing"""
-        return self._data_model.get_properties()
-
-    def _handle_data_changed(self):
-        if self._data_model.has_duplicate_key():
-            self._message_label.setText("Error: there are duplicate property names")
-            self._save_button.setEnabled(False)
-        else:
-            if self._data_model.is_modified():
-                self._message_label.setText("Modified")
-                self._save_button.setEnabled(True)
-            else:
-                self._message_label.setText("No changes")
-                self._save_button.setEnabled(False)
